requirements.txt
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
matplotlib==3.0.3
numpy==1.16.2
opencv-contrib-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
scikit-image==0.14.3
scikit-learn==0.19.2
tensorflow==1.12.0
tqdm==4.31.1